backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ML (NumPy only - ground up implementation)
numpy>=1.26.0
scipy>=1.12.0  # Only for stats functions, not ML

# Gemini
google-generativeai>=0.3.0

# Utilities
httpx>=0.26.0
python-dateutil>=2.8.2

# Dev
pytest>=7.4.0
pytest-asyncio>=0.23.0
